Former-commit-id: db041160f4c13e84dbdd9e15f78f59560daa8eb8 [formerly 70a12a6a500cfb1cf2074a4fdca4c22e29e34947]
Former-commit-id: a8df14104d6492519962b1bcd72bf36d607b4f14
Former-commit-id: 2b3b728007e306190ceb82ce0f3faae864509e0b
Former-commit-id: 31432243b0dfe8527fb7b1135a1b00b4a7652aa1

+++ Lib/pagebot/elements/views/siteview.py
@@ -79,8 +79,11 @@ class SiteView(HtmlView):
                     # Safety check, only run on relative paths
                     assert dstPath.startswith('/tmp/') or not dstPath.startswith('/'), ('Path must be relative: "%s"' % dstPath) 
                     shutil.rmtree(dstPath)
-                # TODO: Fails in Travis.
-                shutil.copytree(resourcePath, dstPath)
+                if os.path.exists(resourcePath):
+                    shutil.copytree(resourcePath, dstPath)
+                elif self.verbose:
+                    print('[%s.build] Resource "%s" does not exist.' % (self.__class__.__name__, resourcePath))
+
 
     def build(self, path=None, pageSelection=None, multiPage=True):
         """
